Rename some file and vars
