+++ src/routers/root.js
@@ -6,16 +6,16 @@ const {
   updateLinkedAccountIDSchema,
   getUserByIDSchema,
   getUserByLinkedAccountIDSchema
-} = require('../schemas/services')
-const { RootController } = require('../services/root')
+} = require('../schemas/services/root')
+const { RootService } = require('../services/root')
 
 module.exports = async (server, opts) => {
   const { mongol } = opts
-  const rootController = new RootController(mongol)
+  const rootService = new RootService(mongol)
 
   server.get('/', { schema: getAllUsersSchema }, async (req, res) => {
     try {
-      let result = await rootController.getAllUsers()
+      let result = await rootService.getAllUsers()
       res.status(200).send(result)
     } catch (err) {
       server.log.error(err.message)
@@ -26,7 +26,7 @@ module.exports = async (server, opts) => {
   server.post('/', { schema: createUserSchema }, async (req, res) => {
     const { username, name, avatar, email, birthday, linkedAccounts } = req.body
     try {
-      const { insertedId } = await rootController.createUser(
+      const { insertedId } = await rootService.createUser(
         username,
         name,
         avatar,
@@ -44,7 +44,7 @@ module.exports = async (server, opts) => {
   server.get('/:id', { schema: getUserByIDSchema }, async (req, res) => {
     const _id = new ObjectID(req.params.id)
     try {
-      let result = await rootController.getUserByID(_id)
+      let result = await rootService.getUserByID(_id)
       res.status(200).send(result)
     } catch (err) {
       server.log.error(err.message)
@@ -58,7 +58,7 @@ module.exports = async (server, opts) => {
     async (req, res) => {
       const { service, id } = req.params
       try {
-        const user = await rootController.getUserByLinkedAccountID(service, id)
+        const user = await rootService.getUserByLinkedAccountID(service, id)
         res.status(200).send(user)
       } catch (err) {
         server.log.error(err.message)
@@ -73,7 +73,7 @@ module.exports = async (server, opts) => {
     async (req, res) => {
       const { service, id, newID } = req.params
       try {
-        await rootController.updateLinkedAccountID(service, id, newID)
+        await rootService.updateLinkedAccountID(service, id, newID)
         res.status(204)
       } catch (err) {
         server.log.error(err.message)
